Week_8/requirements.txt
pandas
google-generativeai
scikit-learn
python-dotenv
streamlit


